% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw_471.R
\name{hw_471}
\alias{hw_471}
\title{STAT471 Homework Format}
\usage{
hw_471(documentclass = c("tufte-handout", "tufte-book"), fig_width = 4,
  fig_height = 2.5, fig_crop = TRUE, dev = "pdf", highlight = "default",
  template = template_resources("hw_471", "template.tex"), ...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{dev}{Graphics device to use for figure output (defaults to pdf)}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{template}{Pandoc template to use for rendering. Pass "default" to use
the rmarkdown package default template; pass \code{NULL} to use pandoc's
built-in template; pass a path to use a custom template that you've
created.  See the documentation on
\href{{http://pandoc.org/README.html}{pandoc online documentation}}
for details on creating custom templates.}

\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
STAT471 Homework Format
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "hw_471", package = "stat471")
}

}

