% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{load_data}
\alias{load_data}
\title{Load a dataset}
\usage{
load_data(x)
}
\arguments{
\item{x}{dataset to load}
}
\value{
dataset
}
\description{
Simple wrapper around utills::data - students often will name their
own datasets 'data' which can create collision problems.
}
\examples{
load_data(sleep)
}

