% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ml_pay}
\alias{ml_pay}
\title{Major League Pay Data - Total}
\format{A data frame with 5 variables:
* `payroll`: total pay up to 2014 in billion dollars
* `avgwin`: average winning percentage for the span of 1998 to 2014
* `p2014`: total pay in 2014
* `X2014`: number of games won in 2014.
* `X2014.pct`: percent of games won in 2014}
\usage{
ml_pay
}
\description{
Formerly: MLPayData_Total.csv. Winning records and the payroll of all
30 ML teams from 1998 to 2014. Note that there are 162 games in each season.
}
\keyword{datasets}

