% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat471-package.R
\docType{package}
\name{stat471}
\alias{stat471}
\alias{stat471-package}
\title{R Package for Modern Data Mining}
\description{
Chris Hua <chua@wharton.upenn.edu>
}
\details{
Usage:
\code{if(!require('devtools')) install.packages('devtools')}
\code{devtools::install_github("stillmatic/stat471")}

In order to create PDF files from your R markdown source, you will also need
to install TeX. Mac users can install MacTeX from
\url{http://tug.org/mactex}; Windows users can install MikTeX
(\url{http://miktex.org/download}) or TeX Live
(\url{http://tug.org/texlive}).
}

