% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tidy_ml_data}
\alias{tidy_ml_data}
\title{Converts MLB payroll data into tidy format}
\usage{
tidy_ml_data()
}
\value{
tidy dataframe with MLB team, year, wins, payroll, winrate
}
\description{
Converts MLB payroll data into tidy format
}
\examples{
head(tidy_ml_data)
}

