% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_471}
\alias{update_471}
\title{Updates this package}
\usage{
update_471()
}
\value{
updates package
}
\description{
Installs package from source
}
\details{
TODO: automate checks against Github version, autoupdate on load
}
\examples{
\dontrun{update_471}
}

